Code split webgl-related code based on offscreen canvas presence via React.lazy
